api/pyproject.toml
[tool.poetry]
name = "sat10am-pages-api"
version = "0.1.0"
description = "A API backend for SAT10AM Pages Project"
authors = ["DoonDoony"]

[tool.poetry.dependencies]
python = "^3.8"
django = "2.1"
djangorestframework = "^3.11.0"
django-cors-headers = "^3.2.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
mypy = "^0.770"
pylint = "^2.5.0"
isort = "^4.3.21"
autoflake = "^1.3.1"
ipython = "^7.13.0"

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
line_length = "79"
multi_line_output = "5"
length_sort = "1"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
